NetHunter Additions

+++ drivers/net/can/usb/peak_usb/pcan_usb_pro.c
@@ -127,7 +127,9 @@ static u8 *pcan_msg_init_empty(struct pcan_usb_pro_msg *pm,
 /*
  * add one record to a message being built
  */
-static int pcan_msg_add_rec(struct pcan_usb_pro_msg *pm, u8 id, ...)
+//static int pcan_msg_add_rec(struct pcan_usb_pro_msg *pm, u8 id, ...)
+//drivers:net:can:usb:peak_usb:pcan_usb_pro.c fix undefined behavior
+static int pcan_msg_add_rec(struct pcan_usb_pro_msg *pm, int id, ...)
 {
 	int len, i;
 	u8 *pc;
